+++ api/management/commands/ingest_ns_directory.py
@@ -2,16 +2,26 @@ import requests
 import xmltodict
 from django.conf import settings
 from django.core.management.base import BaseCommand
+from django.db import transaction
 from requests.auth import HTTPBasicAuth
+from sentry_sdk.crons import monitor
 
 from api.logger import logger
 from api.models import Country, CountryDirectory, CronJob, CronJobStatus
+from main.sentry import SentryMonitor
 
 
 class Command(BaseCommand):
     help = "Add ns contact details"
 
+    @monitor(monitor_slug=SentryMonitor.INGEST_NS_DIRECTORY)
+    @transaction.atomic
     def handle(self, *args, **kwargs):
+        def postprocessor(path, key, value):
+            if key == "@i:nil":
+                return None
+            return key, value
+
         logger.info("Starting NS Contacts")
         url = "https://go-api.ifrc.org/"
         headers = {"accept": "application/xml;q=0.9, */*;q=0.8"}
@@ -33,7 +43,8 @@ class Command(BaseCommand):
             raise Exception("Error querying NationalSocietiesContacts")
 
         added = 0
-        dict_data = xmltodict.parse(response.content)
+        created_country_directory_ids = []
+        dict_data = xmltodict.parse(response.content, postprocessor=postprocessor)
         for data in dict_data["ArrayOfNationalSocietiesContacts"]["NationalSocietiesContacts"]:
             country_name = data["CON_country"] if isinstance(data["CON_country"], str) else None
             if country_name is not None:
@@ -54,7 +65,20 @@ class Command(BaseCommand):
                         "position": data["CON_title"],
                         "country": country,
                     }
-                    CountryDirectory.objects.create(**data)
+                    country_directory, _ = CountryDirectory.objects.get_or_create(
+                        country=country,
+                        first_name__iexact=data["first_name"],
+                        last_name__iexact=data["last_name"],
+                        position__iexact=data["position"],
+                        defaults={
+                            "first_name": data["first_name"],
+                            "last_name": data["last_name"],
+                            "position": data["position"],
+                        },
+                    )
+                    created_country_directory_ids.append(country_directory.pk)
+        # NOTE: Deleting the country directory which are not available in the source
+        CountryDirectory.objects.exclude(id__in=created_country_directory_ids).delete()
         text_to_log = "%s Ns Directory added" % added
         logger.info(text_to_log)
         body = {"name": "ingest_ns_directory", "message": text_to_log, "num_result": added, "status": CronJobStatus.SUCCESSFUL}