+++ databank/management/commands/ingest_databank.py
@@ -3,9 +3,11 @@ import logging
 
 from django.core.management.base import BaseCommand
 from django.utils import timezone
+from sentry_sdk.crons import monitor
 
 from api.models import Country, CronJob, CronJobStatus
 from databank.models import CountryOverview
+from main.sentry import SentryMonitor
 
 from .sources import FDRS, FTS_HPC, INFORM, RELIEFWEB, START_NETWORK, WB
 
@@ -108,6 +110,7 @@ class Command(BaseCommand):
                 }
             )
 
+    @monitor(monitor_slug=SentryMonitor.INGEST_DATABANK)
     def handle(self, *args, **kwargs):
         start = datetime.datetime.now()
         self.load()