+++ databank/management/commands/ingest_hdr.py
@@ -2,8 +2,11 @@ import logging
 
 import requests
 from django.core.management.base import BaseCommand
+from django.db import models, transaction
+from sentry_sdk.crons import monitor
 
-from databank.models import CountryOverview as CO
+from databank.models import CountryOverview
+from main.sentry import SentryMonitor
 
 logger = logging.getLogger(__name__)
 
@@ -11,15 +14,29 @@ logger = logging.getLogger(__name__)
 class Command(BaseCommand):
     help = "Add HDR GII data"
 
+    @monitor(monitor_slug=SentryMonitor.INGEST_HDR)
+    @transaction.atomic
     def handle(self, *args, **kwargs):
-        for overview in CO.objects.all():
-            HDR_API = f"https://api.hdrdata.org/CountryIndicators/filter?country={overview.country.iso3}&year=2021&indicator=gii"
-            hdr_entities = requests.get(HDR_API)
+        overview_qs = CountryOverview.objects.annotate(
+            country_iso3=models.F("country__iso3"),
+        )
+        for overview in overview_qs.all():
+            hdr_entities = requests.get(
+                "https://api.hdrdata.org/CountryIndicators/filter",
+                params={
+                    "country": overview.country_iso3,
+                    "year": 2021,
+                    "indicator": "gii",
+                },
+            )
             if hdr_entities.status_code != 200:
                 continue
             hdr_entities.raise_for_status()
             hdr_entities = hdr_entities.json()
-            if len(hdr_entities):
-                hdr_gii = hdr_entities[0]["value"]
-                overview.hdr_gii = hdr_gii
-                overview.save(update_fields=["hdr_gii"])
+
+            if len(hdr_entities) == 0:
+                continue
+
+            hdr_gii = hdr_entities[0]["value"]
+            overview.hdr_gii = hdr_gii
+            overview.save(update_fields=["hdr_gii"])