+++ deployments/factories/user.py
@@ -1,6 +1,8 @@
 import factory
 from django.contrib.auth import get_user_model
 
+from api.models import Profile
+
 
 class UserFactory(factory.django.DjangoModelFactory):
     class Meta:
@@ -8,3 +10,12 @@ class UserFactory(factory.django.DjangoModelFactory):
 
     username = factory.Sequence(lambda n: "user_%d" % n)
     email = factory.Sequence(lambda n: "user_%d@ifrc.org" % n)
+
+    @factory.post_generation
+    def create_profile(obj, create, extracted, **kwargs):
+        if create:
+            profile = Profile.objects.get(user=obj)
+            profile.limit_access_to_guest = False
+            profile.save(update_fields=["limit_access_to_guest"])
+            # Set new profile to the user object
+            obj.profile = profile