+++ local_units/views.py
@@ -33,6 +33,7 @@ from local_units.serializers import (
     PrivateLocalUnitDetailSerializer,
     PrivateLocalUnitSerializer,
 )
+from main.permissions import DenyGuestUserPermission
 
 
 class PrivateLocalUnitViewSet(viewsets.ModelViewSet):
@@ -47,7 +48,7 @@ class PrivateLocalUnitViewSet(viewsets.ModelViewSet):
         "local_branch_name",
         "english_branch_name",
     )
-    permission_classes = [permissions.IsAuthenticated, IsAuthenticatedForLocalUnit]
+    permission_classes = [permissions.IsAuthenticated, IsAuthenticatedForLocalUnit, DenyGuestUserPermission]
 
     def get_serializer_class(self):
         if self.action == "list":
@@ -63,7 +64,7 @@ class PrivateLocalUnitViewSet(viewsets.ModelViewSet):
         url_path="validate",
         methods=["post"],
         serializer_class=PrivateLocalUnitSerializer,
-        permission_classes=[permissions.IsAuthenticated, ValidateLocalUnitPermission],
+        permission_classes=[permissions.IsAuthenticated, ValidateLocalUnitPermission, DenyGuestUserPermission],
     )
     def get_validate(self, request, pk=None, version=None):
         local_unit = self.get_object()
@@ -135,4 +136,7 @@ class DelegationOfficeListAPIView(ListAPIView):
 class DelegationOfficeDetailAPIView(RetrieveAPIView):
     queryset = DelegationOffice.objects.all()
     serializer_class = DelegationOfficeSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [
+        permissions.IsAuthenticated,
+        DenyGuestUserPermission,
+    ]